Wed, Oct 2, 2024, 1:42 PM -03:00

+++ pdf.c
@@ -15,6 +15,6 @@ int main(int argc, char *argv[])
 
     FILE *f = fopen(filename, "r");
 
-    int buffer[4];
+    BYTE buffer[4];
     fread()
 }
